requirements.txt
fastapi
uvicorn[standard]
ultralytics
Pillow
python-multipart
streamlit
requests
Pillow